--- src/main/java/com/ssafepole/backend/controller/BoardController.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.controller;
-
-public class BoardController {
-
-}
