--- src/main/java/com/ssafepole/backend/controller/PostController.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.controller;
-
-public class PostController {
-
-}
