--- src/main/java/com/ssafepole/backend/controller/UserController.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.controller;
-
-public class UserController {
-
-}
