--- src/main/java/com/ssafepole/backend/domain/Board.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.domain;
-
-public class Board {
-
-}
