--- src/main/java/com/ssafepole/backend/domain/MessageRoom.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.domain;
-
-public class MessageRoom {
-
-}
