--- src/main/java/com/ssafepole/backend/domain/Role.java
@@ -1,6 +1,0 @@
-package com.ssafepole.backend.domain;
-
-public enum Role {
-    USER,
-    ADMIN
-}
