+++ src/main/java/com/ssafepole/backend/domain/board/domain/Board.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.board.domain;
+
+public class Board {
+
+}
