+++ src/main/java/com/ssafepole/backend/domain/board/domain/repository/BoardRepository.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.board.domain.repository;
+
+public interface BoardRepository {
+
+}
