+++ src/main/java/com/ssafepole/backend/domain/board/presentation/BoardController.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.board.presentation;
+
+public class BoardController {
+
+}
