+++ src/main/java/com/ssafepole/backend/domain/board/service/BoardService.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.board.service;
+
+public interface BoardService {
+
+}
