+++ src/main/java/com/ssafepole/backend/domain/comment/domain/Comment.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.comment.domain;
+
+public class Comment {
+
+}
