+++ src/main/java/com/ssafepole/backend/domain/comment/domain/repository/CommentRepository.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.comment.domain.repository;
+
+public interface CommentRepository {
+
+}
