+++ src/main/java/com/ssafepole/backend/domain/comment/presentation/CommentController.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.comment.presentation;
+
+public class CommentController {
+
+}
