+++ src/main/java/com/ssafepole/backend/domain/comment/service/CommentService.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.comment.service;
+
+public interface CommentService {
+
+}
