+++ src/main/java/com/ssafepole/backend/domain/comment/service/CommentServiceImpl.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.comment.service;
+
+public class CommentServiceImpl {
+
+}
