+++ src/main/java/com/ssafepole/backend/domain/message/domain/Message.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.message.domain;
+
+public class Message {
+
+}
