+++ src/main/java/com/ssafepole/backend/domain/message/domain/reopsitory/MessageRepository.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.message.domain.reopsitory;
+
+public interface MessageRepository {
+
+}
