+++ src/main/java/com/ssafepole/backend/domain/message/presentation/MessageController.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.message.presentation;
+
+public class MessageController {
+
+}
