+++ src/main/java/com/ssafepole/backend/domain/message/service/MessageService.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.message.service;
+
+public interface MessageService {
+
+}
