+++ src/main/java/com/ssafepole/backend/domain/message/service/MessageServiceImpl.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.message.service;
+
+public class MessageServiceImpl {
+
+}
