+++ src/main/java/com/ssafepole/backend/domain/messageroom/domain/MessageRoom.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.messageroom.domain;
+
+public class MessageRoom {
+
+}
