+++ src/main/java/com/ssafepole/backend/domain/messageroom/domain/repository/MessageRoomRepository.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.messageroom.domain.repository;
+
+public interface MessageRoomRepository {
+
+}
