+++ src/main/java/com/ssafepole/backend/domain/messageroom/service/MessageRoomServiceImpl.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.messageroom.service;
+
+public class MessageRoomServiceImpl {
+
+}
