+++ src/main/java/com/ssafepole/backend/domain/post/domain/repository/PostRepository.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.post.domain.repository;
+
+public interface PostRepository {
+
+}
