+++ src/main/java/com/ssafepole/backend/domain/post/presentation/PostController.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.post.presentation;
+
+public class PostController {
+
+}
