+++ src/main/java/com/ssafepole/backend/domain/post/service/PostService.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.post.service;
+
+public interface PostService {
+
+}
