+++ src/main/java/com/ssafepole/backend/domain/user/domain/Role.java
@@ -1,0 +1,6 @@
+package com.ssafepole.backend.domain.user.domain;
+
+public enum Role {
+    USER,
+    ADMIN
+}
