+++ src/main/java/com/ssafepole/backend/domain/user/domain/repository/UserRepository.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.user.domain.repository;
+
+public interface UserRepository {
+
+}
