+++ src/main/java/com/ssafepole/backend/domain/user/presentation/UserController.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.user.presentation;
+
+public class UserController {
+
+}
