+++ src/main/java/com/ssafepole/backend/domain/user/service/UserService.java
@@ -1,0 +1,5 @@
+package com.ssafepole.backend.domain.user.service;
+
+public interface UserService {
+
+}
