--- src/main/java/com/ssafepole/backend/repository/BoardRepository.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.repository;
-
-public interface BoardRepository {
-
-}
