--- src/main/java/com/ssafepole/backend/repository/CommentRepository.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.repository;
-
-public interface CommentRepository {
-
-}
