--- src/main/java/com/ssafepole/backend/repository/MessageRepository.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.repository;
-
-public interface MessageRepository {
-
-}
