--- src/main/java/com/ssafepole/backend/repository/PostRepository.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.repository;
-
-public interface PostRepository {
-
-}
