--- src/main/java/com/ssafepole/backend/service/board/BoardService.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.service.board;
-
-public interface BoardService {
-
-}
