--- src/main/java/com/ssafepole/backend/service/board/BoardServiceImpl.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.service.board;
-
-public class BoardServiceImpl {
-
-}
