--- src/main/java/com/ssafepole/backend/service/comment/CommentServiceImpl.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.service.comment;
-
-public class CommentServiceImpl {
-
-}
