--- src/main/java/com/ssafepole/backend/service/message/MessageService.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.service.message;
-
-public interface MessageService {
-
-}
