--- src/main/java/com/ssafepole/backend/service/message/MessageServiceImpl.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.service.message;
-
-public class MessageServiceImpl {
-
-}
