--- src/main/java/com/ssafepole/backend/service/messageroom/MessageRoomService.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.service.messageroom;
-
-public interface MessageRoomService {
-
-}
