--- src/main/java/com/ssafepole/backend/service/messageroom/MessageRoomServiceImpl.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.service.messageroom;
-
-public class MessageRoomServiceImpl {
-
-}
