--- src/main/java/com/ssafepole/backend/service/post/PostService.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.service.post;
-
-public interface PostService {
-
-}
