--- src/main/java/com/ssafepole/backend/service/post/PostServiceImpl.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.service.post;
-
-public class PostServiceImpl {
-
-}
