--- src/main/java/com/ssafepole/backend/service/user/UserService.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.service.user;
-
-public interface UserService {
-
-}
