--- src/main/java/com/ssafepole/backend/service/user/UserServiceImpl.java
@@ -1,5 +1,0 @@
-package com.ssafepole.backend.service.user;
-
-public class UserServiceImpl {
-
-}
